refactor: custom hooks for simulations
- Wrap `getSimulateInput` and `getVariablesSimulated` in memo-ised custom hooks.`
- run dataset data-fetching requests as soon as the dataset ID is ready, not the dataset itself.
